ready for web interface (#289)
